+++ src/main/java/com/hibit2/hibit2/auth/application/OAuthUri.java
@@ -1,4 +1,4 @@
-package com.hibit2.hibit2.auth.support;
+package com.hibit2.hibit2.auth.application;
