+++ src/test/java/com/hibit2/hibit2/common/config/ExternalApiConfig.java
@@ -1,7 +1,7 @@
 package com.hibit2.hibit2.common.config;
 
-import com.hibit2.hibit2.auth.support.OAuthClient;
-import com.hibit2.hibit2.auth.support.OAuthUri;
+import com.hibit2.hibit2.auth.application.OAuthClient;
+import com.hibit2.hibit2.auth.application.OAuthUri;
 import com.hibit2.hibit2.infrastructure.oauth.client.StubOAuthClient;
 import com.hibit2.hibit2.infrastructure.oauth.uri.StubOAuthUri;
 import org.springframework.boot.test.context.TestConfiguration;
